requirements.txt
numpy>=1.26
pandas>=2.2
scikit-learn>=1.5
xgboost>=2.1
torch>=2.4
matplotlib>=3.9
seaborn>=0.13
shap>=0.46
pyyaml>=6.0
pydantic>=2.0
lime>=0.2
